Updating models and controllers for orders
Changes have been made to the CartsController.cs controller to return the cart object instead of throwing an exception when there is no cart.
New DbSets for Order and OrderItem have been added to TameShopDbContext.cs with definitions of relationships between entities.
The data type for the Price property in the Animal entity has been changed from double to decimal in UserDbContextModelSnapshot.cs, and the indexes and table names for Order and OrderItem have been updated.
Added virtual navigation properties to the Animal and CartItem models.
Added the User property to Order and the Order property to OrderItem to link to the order.
Created migrations to initialize the new models and their relationships, as well as to rename tables and update foreign keys.

+++ Models/Animal.cs
@@ -6,7 +6,7 @@ namespace TameShop.Models
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public double Price { get; set; } = 1.99;
+        public decimal Price { get; set; } = 1.99M;
         public string Description { get; set; } = string.Empty;
         public string ImageUrl { get; set; } = string.Empty;
         public string Category { get; set; } = string.Empty;
