+++ ViewModels/AnimalDTO.cs
@@ -9,7 +9,7 @@ namespace TameShop.ViewModels
         public string Name { get; set; } = string.Empty;
         [Required]
         [Range(0.01, double.MaxValue, ErrorMessage = "Price must be greater than zero.")]
-        public double Price { get; set; } = 1.99;
+        public decimal Price { get; set; } = 1.99M;
         public string Description { get; set; } = string.Empty;
         public string ImageUrl { get; set; } = string.Empty;
         public string Category { get; set; } = string.Empty;
